+++ fusionlab/layers/__init__.py
@@ -1,7 +1,7 @@
 from fusionlab.configs import BACKEND
 if BACKEND == 'torch':
-    from .squeeze_excitation import SEModule
+    from .squeeze_excitation.se import SEModule
 elif BACKEND == 'tf':
-    from .squeeze_excitation import TFSEModule
+    from .squeeze_excitation.tfse import TFSEModule
 else:
     print('backend not supported!!!')
