+++ scripts/build_pip.sh
@@ -2,7 +2,7 @@
 # ref: https://datainpoint.substack.com/p/c0a
 
 python ../setup.py sdist bdist_wheel
-python -m twine upload ../dist/*
+python -m twine upload dist/*
 
 # build
 python setup.py sdist bdist_wheel
